mtd: nand: atmel: Set nfc->clk to NULL when the clock is missing in the DT
The nfc->clk seems to be optional. Instead leaving the nfc->clk field to
an error pointer when the clk is missing, assign it to NULL to avoid
checking the pointer validity each time we want to enable/disable
the clk (the clk framework already nicely handles the !clk case).

Signed-off-by: Boris Brezillon <boris.brezillon@free-electrons.com>