clk: at91: utmi: set the mainck rate
By default, it is assumed that the UTMI clock is generated from a 12 MHz
reference clock (MAINCK). If it's not the case, the FREQ field of the
SFR_UTMICKTRIM has to be updated to generate the UTMI clock in the
proper way.

The UTMI clock has a fixed rate of 480 MHz. In fact, there is no multiplier
we can configure. The multiplier is managed internally, depending on the
reference clock frequency, to achieve the target of 480 MHz.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>